ebay/requirements.txt
azure-functions==1.7.2
certifi==2021.5.30
ebaysdk==2.2.0